requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
certifi==2018.10.15
chardet==3.0.4
conekta==2.4.0
Django==1.11.13
django-appconf==1.0.2
django-filter==1.1.0
django-image-cropping==1.2.0
djangorestframework==3.6.4
easy-thumbnails==2.5
haversine==0.4.5
idna==2.7
kombu==3.0.37
Markdown==2.6.11
nose==1.3.7
Pillow==5.1.0
python-dateutil==2.7.5
pytz==2018.4
redis==2.10.3
requests==2.20.0
simplejson==3.16.0
six==1.12.0
urllib3==1.24
XlsxWriter==1.0.5
